% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BL1.property.R
\docType{data}
\name{BL1.property}
\alias{BL1.property}
\title{Genelist for BL1.property}
\format{A dataframe with 2 columns, and 473 rows
 describe{
   item Gene.Symbol ID of transcript (HGNC where available)
   item Direction Direction of expression (up/down)
}}
\source{
Identified in related publication
}
\usage{
BL1.property
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\keyword{datasets}

