% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.stroma.property.R
\docType{data}
\name{D.stroma.property}
\alias{D.stroma.property}
\title{\name{D.stroma.property}
\alias{D.stroma.property}
\docType{data}
\title {Genelist for D.stroma.property}}
\format{An object of class \code{data.frame} with 297 rows and 2 columns.}
\usage{
D.stroma.property
}
\description{
\description{This genelist is used to assign the related property using the assign.property function}
\usage{data('D.stroma.property')}
\format{ A dataframe 297 observations on the following 2 variables:
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}
}
\source Identified in related publication
\examples{data('D.stroma.property')}
}
\keyword{datasets}

