% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.roi.regions.R
\name{define.roi.regions}
\alias{define.roi.regions}
\title{Function to assign samples as low intermediate or high based on ROI}
\usage{
define.roi.regions(bresat, random.rank.counts, middle.range = 0.95)
}
\arguments{
\item{bresat}{Full return data returned from the sig.ranksum function.}

\item{random.rank.counts}{ROI region returned from random.ranks function.}

\item{middle.range}{The ROI range).}
}
\description{
This is a function to samples as low intermediate or high based on ROI
}

