% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.stroma.property.R
\docType{data}
\name{E.stroma.property}
\alias{E.stroma.property}
\title{Genelist for E.stroma.property}
\format{A dataframe with 2 columns, and 27 rows
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}}
\source{
Identified in related publication
}
\usage{
E.stroma.property
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\keyword{datasets}

