% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IM.property.R
\docType{data}
\name{IM.property}
\alias{IM.property}
\title{\name{IM.property}
\alias{IM.property}
\docType{data}
\title {Genelist for IM.property}}
\format{An object of class \code{data.frame} with 705 rows and 2 columns.}
\usage{
IM.property
}
\description{
\description{This genelist is used to assign the related property using the assign.property function}
\usage{data('IM.property')}
\format{ A dataframe 705 observations on the following 2 variables:
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}
}
\source Identified in related publication
\examples{data('IM.property')}
}
\keyword{datasets}

