% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAR.property.R
\docType{data}
\name{LAR.property}
\alias{LAR.property}
\title{\name{LAR.property}
\alias{LAR.property}
\docType{data}
\title {Genelist for LAR.property}}
\format{An object of class \code{data.frame} with 1205 rows and 2 columns.}
\usage{
LAR.property
}
\description{
\description{This genelist is used to assign the related property using the assign.property function}
\usage{data('LAR.property')}
\format{ A dataframe 1205 observations on the following 2 variables:
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}
}
\source Identified in related publication
\examples{data('LAR.property')}
}
\keyword{datasets}

