% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.property.R
\docType{data}
\name{M.property}
\alias{M.property}
\title{\name{M.property}
\alias{M.property}
\docType{data}
\title {Genelist for M.property}}
\format{An object of class \code{data.frame} with 710 rows and 2 columns.}
\usage{
M.property
}
\description{
\description{This genelist is used to assign the related property using the assign.property function}
\usage{data('M.property')}
\format{ A dataframe 710 observations on the following 2 variables:
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}
}
\source Identified in related publication
\examples{data('M.property')}
}
\keyword{datasets}

