% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSL.property.R
\docType{data}
\name{MSL.property}
\alias{MSL.property}
\title{\name{MSL.property}
\alias{MSL.property}
\docType{data}
\title {Genelist for MSL.property}}
\format{An object of class \code{data.frame} with 1105 rows and 2 columns.}
\usage{
MSL.property
}
\description{
\description{This genelist is used to assign the related property using the assign.property function}
\usage{data('MSL.property')}
\format{ A dataframe 1105 observations on the following 2 variables:
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}
}
\source Identified in related publication
\examples{data('MSL.property')}
}
\keyword{datasets}

