% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.ranksum.R
\name{sig.ranksum}
\alias{sig.ranksum}
\title{Function to assign ranksums to an expression matrix}
\usage{
sig.ranksum(exprdata, up = NULL, dn = NULL, ns = NULL,
  full.return = FALSE)
}
\arguments{
\item{exprdata}{An expression matrix. Rows correspont to genes, columns to samples. Genes with multiple probes should be collapsed to a single row e.g. using most variable probe.}

\item{up}{A vector specifying rows of genes in the up direction.}

\item{dn}{A vector specifying rows of genes in the down direction.}

\item{ns}{A vector specifying rows of genes with an unspecified direction. These will be split using the pam function from the cluster package.}

\item{full.return}{A boolean specifying if the full result should be returned (T) or only the ranksum values (F).}
}
\description{
This is a function to assign ranksums to a dataset
}

