% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T.stroma.property.R
\docType{data}
\name{T.stroma.property}
\alias{T.stroma.property}
\title{\name{T.stroma.property}
\alias{T.stroma.property}
\docType{data}
\title {Genelist for T.stroma.property}}
\format{An object of class \code{data.frame} with 458 rows and 2 columns.}
\usage{
T.stroma.property
}
\description{
\description{This genelist is used to assign the related property using the assign.property function}
\usage{data('T.stroma.property')}
\format{ A dataframe 458 observations on the following 2 variables:
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}
}
\source Identified in related publication
\examples{data('T.stroma.property')}
}
\keyword{datasets}

