% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.property.R
\name{assign.properties}
\alias{assign.properties}
\title{Function to assign properties to an expression matrix}
\usage{
assign.properties(ESet, geneID.column = 1, genelists = c("Stroma4",
  "TNBCType"), n = 1000, seed = 123456, mc.cores = snowWorkers(),
  var.method = function(x) rowIQRs(x, na.rm = TRUE))
}
\arguments{
\item{ESet}{An ExpressionSet object. Rows correspond to genes, columns to samples. If there are genes with multiple probes, they will be collapsed to a single row using the function defined in var.method.}

\item{geneID.column}{Integer or column name corresponding to column in featureData table corresponding to HGNC ID.}

\item{genelists}{A vector with either Stroma4, TNBCType, or both to specify which genelists to use.}

\item{n}{An integer value specifying the number of random samples to generate.}

\item{seed}{An integer value specifying a seed for the random ranks function. Default value is 123456.}

\item{mc.cores}{An integer specifying how many cores to use. Defaults to use the function snowWorkers().}

\item{var.method}{Function for assessing variance to collapse probes. Default is IQR}
}
\value{
The function returns a the ExpressionSet with the properties added to the pData table.
}
\description{
\code{assign.properties} is a function to assign stromal property and Lehmann generative property levels to a TNBC dataset. Given an expressionset with a gene expression matrix, and a a list of HGNC IDs, this function will estimate the status of samples for either the Lehmann generative properties, the stromal properties, or both at the same time. (Saleh et al. , under review at Cancer Research)
}
\details{
We defined a method for estimating stromal and Lehmann generative properties in a gene expression dataset (either LCM stroma or whole tumor breast cancer data). The \code{assign.properties} function extracts the expression data from the expression set using the exprs() function. The HGNC ID annotations for the dataset are required as this is what the function uses to identify genes to estimate property levels. The HGNC IDs should be in the featureData of the expressionset, and the corresponding column ID should be passed to the geneID.column argument.  The function can be used to estimate the stromal properties, the  Lehmann generative properties or both simultaneously depending on whether 'STROMA4' (stromal properties), 'TNBCTYPE' (Lehmann generative properties), or both are passed to the genelists argument. This function will then return property assignments.
}
\examples{
library(breastCancerMAINZ)
data(mainz, package='breastCancerMAINZ')
all.properties <- assign.properties(ESet=mainz, geneID.column='Gene.symbol',
	genelists=c('Stroma4', 'TNBCType'), n=10)
}
