% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BL2.property.R
\docType{data}
\name{BL2.property}
\alias{BL2.property}
\title{Genelist for BL2.property}
\format{A dataframe with 2 columns, and 291 rows
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}}
\source{
Identified in related publication
}
\usage{
BL2.property
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\keyword{datasets}

