% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BL2.property.R
\docType{data}
\name{BL2.property}
\alias{BL2.property}
\title{Genelist for BL2.property}
\format{An object of class \code{data.frame} with 291 rows and 2 columns.
\itemize{
  \item{Gene.Symbol: }{ID of transcript (HGNC where available)}
  \item{Direction: }{Direction of expression (up/down)}
}}
\source{
Identified in related publication
}
\usage{
data('BL2.property')
}
\value{
An object of class \code{data.frame} with 291 rows and 2 columns.
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\examples{
data('BL2.property')
}
\keyword{List}

