% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.stroma.property.R
\docType{data}
\name{D.stroma.property}
\alias{D.stroma.property}
\title{Genelist for D.stroma.property}
\format{A dataframe with 2 columns, and 297 rows
 describe{
   item Gene.Symbol ID of transcript (HGNC where available)
   item Direction Direction of expression (up/down)
}}
\source{
Identified in related publication
}
\usage{
D.stroma.property
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\keyword{datasets}

