% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.stroma.property.R
\docType{data}
\name{D.stroma.property}
\alias{D.stroma.property}
\title{Genelist for D.stroma.property}
\format{An object of class \code{data.frame} with 297 rows and 2 columns.
\itemize{
  \item{Gene.Symbol: }{ID of transcript (HGNC where available)}
  \item{Direction: }{Direction of expression (up/down)}
}}
\source{
Identified in related publication
}
\usage{
data('D.stroma.property')
}
\value{
An object of class \code{data.frame} with 297 rows and 2 columns.
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\examples{
data('D.stroma.property')
}
\keyword{List}

