% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAR.property.R
\docType{data}
\name{LAR.property}
\alias{LAR.property}
\title{Genelist for LAR.property}
\format{A dataframe with 2 columns, and 1205 rows
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}}
\source{
Identified in related publication
}
\usage{
LAR.property
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\keyword{datasets}

