% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.property.R
\docType{data}
\name{M.property}
\alias{M.property}
\title{Genelist for M.property}
\format{A dataframe with 2 columns, and 710 rows
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}}
\source{
Identified in related publication
}
\usage{
M.property
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\keyword{datasets}

