% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSL.property.R
\docType{data}
\name{MSL.property}
\alias{MSL.property}
\title{Genelist for MSL.property}
\format{A dataframe with 2 columns, and 1105 rows
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}}
\source{
Identified in related publication
}
\usage{
MSL.property
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\keyword{datasets}

