% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T.stroma.property.R
\docType{data}
\name{T.stroma.property}
\alias{T.stroma.property}
\title{Genelist for T.stroma.property}
\format{A dataframe with 2 columns, and 458 rows
\describe{
  \item{Gene.Symbol}{ID of transcript (HGNC where available)}
  \item{Direction}{Direction of expression (up/down)}
}}
\source{
Identified in related publication
}
\usage{
T.stroma.property
}
\description{
This genelist is used to assign the related property using the assign.property function
}
\keyword{datasets}

